consumer/requirements.txt
apache-beam[gcp]==2.50.0
google-cloud-bigquery==3.11.0
google-cloud-pubsub==2.18.0
